chore: renamed utils to lib
